+++ pages/wallet/index.tsx
