--- src/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "shopstr",
-  "version": "0.1.0",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@cashu/cashu-ts": "^0.8.0-rc.4",
-    "@heroicons/react": "^2.0.18",
-    "@types/node": "18.15.11",
-    "autoprefixer": "10.4.14",
-    "axios": "^1.3.5",
-    "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
-    "next": "^13.3.4",
-    "nostr-tools": "^1.14.2",
-    "postcss": "8.4.21",
-    "qrcode": "^1.5.3",
-    "react": "18.2.0",
-    "tailwindcss": "3.3.1",
-    "uuid": "^9.0.0",
-    "websocket-polyfill": "^0.0.3"
-  },
-  "devDependencies": {
-    "@types/react": "^18.0.38",
-    "@types/react-dom": "^18.0.11",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
-  }
-}
