+++ utils/getRelays.ts
@@ -1,0 +1,3 @@
+export default function getRelays(): string[] {
+    return JSON.parse(localStorage.getItem('relays') || '["wss://relay.damus.io"]');
+}
